+++ runners/media-inspector.c
@@ -31,7 +31,7 @@
 #include <libavutil/rational.h>
 #include <libavutil/samplefmt.h>
 
-#define RUNNER_VERSION "1.4.0"
+#define RUNNER_VERSION "1.5.0"
 #define REPORT_SCHEMA_VERSION 1
 
 typedef struct RunnerOptions {